# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ClearActivityRequest": {
        "properties": {
          "older_than_days": {
            "title": "Older Than Days",
            "type": "integer"
          }
        },
        "required": [
          "older_than_days"
        ],
        "title": "ClearActivityRequest",
        "type": "object"
      },
      "ClearableActivityResponse": {
        "properties": {
          "entries": {
            "title": "Entries",
            "type": "integer"
          },
          "older_than_days": {
            "title": "Older Than Days",
            "type": "integer"
          }
        },
        "required": [
          "older_than_days",
          "entries"
        ],
        "title": "ClearableActivityResponse",
        "type": "object"
      },
      "ClearedActivityResponse": {
        "properties": {
          "database": {
            "$ref": "#/components/schemas/DatabaseUsageResponse"
          },
          "removed": {
            "title": "Removed",
            "type": "integer"
          }
        },
        "required": [
          "removed",
          "database"
        ],
        "title": "ClearedActivityResponse",
        "type": "object"
      },
...
      "DatabaseUsageResponse": {
        "properties": {
          "activity_entries": {
            "title": "Activity Entries",
            "type": "integer"
          },
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "oldest_activity_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Activity At"
          },
          "reclaimable_bytes": {
            "title": "Reclaimable Bytes",
            "type": "integer"
          }
        },
        "required": [
          "bytes",
          "reclaimable_bytes",
          "activity_entries",
          "oldest_activity_at"
        ],
        "title": "DatabaseUsageResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LogUsageResponse": {
        "properties": {
          "bytes": {
            "title": "Bytes",
            "type": "integer"
          },
          "files": {
            "title": "Files",
            "type": "integer"
          },
          "newest_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newest At"
          },
          "oldest_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest At"
          }
        },
        "required": [
          "bytes",
          "files",
          "oldest_at",
          "newest_at"
        ],
        "title": "LogUsageResponse",
        "type": "object"
      },
...
      "PasswordRequest": {
        "properties": {
          "password": {
            "maxLength": 256,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "PasswordRequest",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          }
        },
        "required": [
          "authenticated"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "StorageResponse": {
        "properties": {
          "activity_ages": {
            "items": {
              "type": "integer"
            },
            "title": "Activity Ages",
            "type": "array"
          },
          "database": {
            "$ref": "#/components/schemas/DatabaseUsageResponse"
          },
          "logs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LogUsageResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "database",
          "logs",
          "activity_ages"
        ],
        "title": "StorageResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Calendar Ghost",
    "version": "0.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session": {
      "delete": {
        "operationId": "log_out_api_v1_session_delete",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Out"
      },
      "get": {
        "operationId": "session_status_api_v1_session_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Status"
      },
      "post": {
        "operationId": "log_in_api_v1_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log In"
      }
    },
...
    "/api/v1/storage": {
      "get": {
        "operationId": "storage_usage_api_v1_storage_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Storage Usage"
      }
    },
    "/api/v1/storage/activity": {
      "get": {
        "operationId": "clearable_activity_api_v1_storage_activity_get",
        "parameters": [
          {
            "in": "query",
            "name": "older_than_days",
            "required": true,
            "schema": {
              "title": "Older Than Days",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearableActivityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clearable Activity"
      }
    },
    "/api/v1/storage/activity/clear": {
      "post": {
        "operationId": "clear_activity_api_v1_storage_activity_clear_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearActivityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearedActivityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Activity"
      }
    },
    "/api/v1/storage/logs": {
      "delete": {
        "operationId": "purge_logs_api_v1_storage_logs_delete",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge Logs"
      },
      "get": {
        "operationId": "download_logs_api_v1_storage_logs_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "calendar_sync_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar Sync Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Logs"
      }
    },
...
  }
}